requirements.txt
apscheduler==3.11.2
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
oauthlib==3.3.1
outcome==1.3.0.post0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.3.1
pysocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing-extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
websocket-client==1.9.0
wsproto==1.3.2
